+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.12.23
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.20
+  - antd-management-fast-common@2.6.133
+
 ## 2.12.22
 
 ### Patch Changes